fix: docker compose templates

+++ scripts/init.sh
@@ -5,32 +5,30 @@ if [ ! -f .env ]; then
     echo ".env file created"
 fi
 
+read -p "Enter your domain name (default: localhost): " domain
+domain=${domain:-localhost}
+echo -e "\033[36mSetting domain to: $domain\033[0m"
 
-if [ ! -f compose.yml ]; then
-    echo -e "\033[36mcompose.yml not found, creating it from template...\033[0m"
-    cp compose.development.yml.template compose.development.yml
-    cp compose.production.yml.template compose.production.yml
-    echo -e "\033[32mCreated compose.development.yml and compose.production.yml from templates.\033[0m"
+if [ ! -f compose.development.yml ] && [ ! -f compose.production.yml ]; then
+    echo -e "\033[36mcompose.development.yml and compose.production.yml not found, creating them from templates...\033[0m"
     
-    read -p "Enter your domain name (default: localhost): " domain
-    domain=${domain:-localhost}
-    echo -e "\033[36mSetting domain to: $domain\033[0m"
-
-    cp caddy/Caddyfile.template caddy/Caddyfile
-    echo -e "\033[36mGenerating Caddy configuration files...\033[0m"
-    
-    # Create .env file with the domain
     if [ "$domain" = "localhost" ]; then
+        # Development branch
+        cp compose.development.yml.template compose.development.yml
         echo "SITE_ADDRESS=:8080" >> .env
-        # Add the domain to the Caddyfile
+        cp caddy/Caddyfile.template caddy/Caddyfile
         awk '{gsub(/{SITE_ADDRESS}/,":8080")}1' caddy/Caddyfile > caddy/Caddyfile.tmp && mv caddy/Caddyfile.tmp caddy/Caddyfile
-        echo -e "\033[36mAdded SITE_ADDRESS=:80 to .env (localhost mode)\033[0m"
+        echo -e "\033[36mAdded SITE_ADDRESS=:8080 to .env (development mode)\033[0m"
     else
+        # Production branch
+        cp compose.production.yml.template compose.production.yml
         echo "SITE_ADDRESS=${domain}" >> .env
-        # Add the domain to the Caddyfile
+        cp caddy/Caddyfile.template caddy/Caddyfile
         awk '{gsub(/{SITE_ADDRESS}/,"'"$domain"'")}1' caddy/Caddyfile > caddy/Caddyfile.tmp && mv caddy/Caddyfile.tmp caddy/Caddyfile
         echo -e "\033[36mAdded SITE_ADDRESS=${domain} to .env (production mode)\033[0m"
     fi
+    
+    echo -e "\033[32mCreated compose files from templates.\033[0m"
     
     # Create necessary directories
     mkdir -p caddy/data
@@ -42,5 +40,5 @@ if [ ! -f compose.yml ]; then
 
     echo -e "\033[32mConfiguration complete! You can now run 'make d-up' to start the servers.\033[0m"
 else
-    echo -e "\033[33mcompose.yml already exists. If you want to reset it, delete it first and run this command again.\033[0m"
+    echo -e "\033[33mcompose.development.yml and compose.production.yml already exist. If you want to reset them, delete them first and run this command again.\033[0m"
 fi
